Date Formatting
I've added dateFormatter.js, changed the files index.js, index.ejs, article.ejs and styles.css for displaying differently the dates in the index.ejs page depending if we are in mobile display or tablet / desktop display. Also the dates will show correctly as it should when we are viewing an article.